validate request parameters

+++ src/controllers/ImagesController.ts
@@ -1,12 +1,14 @@
-import { OK, NO_CONTENT, CREATED } from 'http-status-codes';
+import { OK, NO_CONTENT, CREATED, BAD_REQUEST } from 'http-status-codes';
 import { Request, Response } from 'express';
 
 import { Controller, Middleware, Get, Post, Delete } from '@overnightjs/core';
 import { Logger } from '@overnightjs/logger';
 
 import upload from '../middlewares/upload';
-
 import ImagesRepository from '../repositories/ImagesRepository';
+
+import validateSearchPattern from '../validators/validateSearchPattern';
+import validateFilename from '../validators/validateFilename';
 
 @Controller('api/images')
 export default class {
@@ -24,8 +26,16 @@ export default class {
     const { pattern } = req.query;
 
     if (typeof (pattern) === 'string') {
-      Logger.Info(`Searching for '${pattern}'`, true);
-      return res.status(OK).json(await this.repository.search(pattern));
+      const validation = validateSearchPattern(pattern);
+
+      if (validation.isValid()) {
+        Logger.Info(`Searching for '${pattern}'`, true);
+        return res.status(OK).json(await this.repository.search(pattern));
+      }
+      else {
+        Logger.Warn(`Search pattern with error '${validation.errorsSummary()}'`, true);
+        return res.status(BAD_REQUEST).json(validation.errorsSummary());
+      }
     }
 
     return res.status(OK).json([]);
@@ -47,9 +57,16 @@ export default class {
   private async delete(req: Request, res: Response) {
     const { friendlyName } = req.params;
 
-    Logger.Info(`Deleting file: ${friendlyName}`, true);
-    await this.repository.remove(friendlyName);
+    const validation = validateFilename(friendlyName);
+    if (validation.isValid()) {
+      Logger.Info(`Deleting file: ${friendlyName}`, true);
+      await this.repository.remove(friendlyName);
 
-    return res.status(NO_CONTENT).json();
+      return res.status(NO_CONTENT).json();
+    }
+    else {
+      Logger.Warn(`friendlyName with error '${validation.errorsSummary()}'`, true);
+      return res.status(BAD_REQUEST).json(validation.errorsSummary());
+    }
   }
 }
